Repo update
Among other things:
- Fear fixed
- Loggable simplified
- L2FortSiegeGuardAI on_intention_attack fixed
- Added EventType::OnCreatureSee
- Fix for Q00325_GrimCollector
- Replaced Scripts::DrChaos with Scripts::GiganticGolem
- Implemented condition::PlayerTvTEvent
- Removed @item_id from L2ItemInstance
- Removed more overflow checks
-

+++ src/network/packets/server/ex_br_premium_state.cr
@@ -1,5 +1,5 @@
 class Packets::Outgoing::ExBrPremiumState < GameServerPacket
-  initializer id : Int32, state : Int32
+  initializer id : Int32, state : Int8
 
   private def write_impl
     c 0xfe
